에러페이지 만들기, 인터셉터, 권한처리

+++ src/main/java/shop/mtcoding/blog/board/BoardController.java
@@ -8,6 +8,8 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import shop.mtcoding.blog._core.errs.exception.Exception403;
+import shop.mtcoding.blog._core.errs.exception.Exception404;
 import shop.mtcoding.blog.user.User;
 
 import java.util.List;
@@ -20,7 +22,7 @@ public class BoardController {
     private final HttpSession session;
 
     @PostMapping("/board/save")
-    public String save(BoardRequest.SaveDTO requestDTO){
+    public String save(BoardRequest.SaveDTO requestDTO) {
         User sessionUser = (User) session.getAttribute("sessionUser");
         //권한 체크는 생략
 
@@ -31,30 +33,46 @@ public class BoardController {
 
 
     @PostMapping("/board/{id}/update")
-    public String update(@PathVariable Integer id, BoardRequest.SaveDTO requestDTO){
+    public String update(@PathVariable Integer id, BoardRequest.SaveDTO requestDTO) {
+        User sessionUser = (User) session.getAttribute("sessionUser");
+        Board board = boardRepository.findById(id);
+
+        if (sessionUser.getId() != board.getUser().getId()) {
+            throw new Exception403("게시글을 수정할 권한이 없습니다");
+        }
+
         boardRepository.updateById(id, requestDTO.getTitle(), requestDTO.getContent());
-        return "redirect:/board/"+id;
+        return "redirect:/board/" + id;
     }
 
     @GetMapping("/board/{id}/update-form")
-    public String updateForm(@PathVariable Integer id, HttpServletRequest request){
+    public String updateForm(@PathVariable Integer id, HttpServletRequest request) {
         Board board = boardRepository.findById(id);
-        request.setAttribute("board", board);
 
+        if (board == null) {
+            throw new Exception404("해당 게시글을 찾을 수 없습니다");
+        }
+
+        request.setAttribute("board", board);
         return "board/update-form";
     }
 
     @PostMapping("/board/{id}/delete")
-    public String delete(@PathVariable Integer id){
-        //권한, 인증 체크 지금은 생략
+    public String delete(@PathVariable Integer id) {
+        User sessionUser = (User) session.getAttribute("sessionUser");
+        Board board = boardRepository.findById(id);
+
+        if (sessionUser.getId() != board.getUser().getId()) {
+            throw new Exception403("게시글을 삭제할 권한이 없습니다");
+        }
+
         boardRepository.deleteById(id);
 
         return "redirect:/";
     }
 
 
-
-    @GetMapping("/" )
+    @GetMapping("/")
     public String index(HttpServletRequest request) {
         List<Board> boardList = boardRepository.findAll();
         request.setAttribute("boardList", boardList);
@@ -68,7 +86,18 @@ public class BoardController {
 
     @GetMapping("/board/{id}")
     public String detail(@PathVariable Integer id, HttpServletRequest request) {
+        User sessionUser = (User) session.getAttribute("sessionUser");
         Board board = boardRepository.findByIdJoinUser(id);
+
+        // 로그인을 하고 게시글의 주인이면 isOwne가 true가 된다 !
+        boolean isOwner = false;
+        if (sessionUser != null) {
+            if (sessionUser.getId() == board.getUser().getId()) {
+                isOwner = true;
+            }
+        }
+
+        request.setAttribute("isOwner", isOwner);
         request.setAttribute("board", board);
         return "board/detail";
     }
